+++ PhyGen.Application/MatrixSections/Commands/UpdateMatrixSectionCommand.cs
@@ -12,7 +12,7 @@ namespace PhyGen.Application.MatrixSections.Commands
         public Guid Id { get; set; }
         public Guid MatrixId { get; set; }
         public string Title { get; set; } = string.Empty;
-        public double? Score { get; set; }
+        public float? Score { get; set; }
         public string? Description { get; set; }
     }
 }
